# api/swagger/gateway/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Платежный шлюз для E-commerce системы",
        "contact": {
            "name": "Курмыза Павел",
            "email": "tmrrwnxtsn@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/favorites/{operation_type}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Избранное"
                ],
                "summary": "Добавить платежную систему в избранные способы оплаты (выплаты)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Тип транзакции",
                        "name": "operation_type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Тело запроса",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.favoritesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "schema": {
                            "$ref": "#/definitions/v1.favoritesResponse"
                        }
                    },
                    "default": {
                        "description": "Ответ с ошибкой",
                        "schema": {
                            "$ref": "#/definitions/v1.errorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Избранное"
                ],
                "summary": "Удалить платежную систему из избранных способов оплаты (выплаты)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Тип транзакции",
                        "name": "operation_type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Тело запроса",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.favoritesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "schema": {
                            "$ref": "#/definitions/v1.favoritesResponse"
                        }
                    },
                    "default": {
                        "description": "Ответ с ошибкой",
                        "schema": {
                            "$ref": "#/definitions/v1.errorResponse"
                        }
                    }
                }
            }
        },
...
        "/payment/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Платежи"
                ],
                "summary": "Создать запрос на пополнение баланса",
                "parameters": [
                    {
                        "description": "Тело запроса",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.paymentCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "schema": {
                            "$ref": "#/definitions/v1.paymentCreateResponse"
                        }
                    },
                    "default": {
                        "description": "Ответ с ошибкой",
                        "schema": {
                            "$ref": "#/definitions/v1.errorResponse"
                        }
                    }
                }
            }
        },
        "/payment/methods": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Платежи"
                ],
                "summary": "Получить список способов для пополнения баланса",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор клиента",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Идентификатор сессии клиента",
                        "name": "session_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Валюта платежа в соответствии со стандартом ISO 4217",
                        "name": "currency",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Код языка, обозначение по RFC 5646",
                        "name": "lang_code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "schema": {
                            "$ref": "#/definitions/v1.paymentMethodsResponse"
                        }
                    },
                    "default": {
                        "description": "Ответ с ошибкой",
                        "schema": {
                            "$ref": "#/definitions/v1.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "v1.commission": {
            "type": "object",
            "required": [
                "type"
            ],
            "properties": {
                "absolute": {
                    "description": "Значение комиссии (передается, если тип комиссии, \"type\", равен \"fixed\" или \"combined\")",
                    "type": "number",
                    "example": 10
                },
                "caption": {
                    "description": "Текстовая репрезентация комиссии (передается, если тип комиссии, \"type\", равен \"text\")",
                    "type": "string",
                    "example": "Комиссия взимается провайдером"
                },
                "currency": {
                    "description": "Код валюты комиссии (передается, если тип комиссии, \"type\", равен \"fixed\" или \"combined\")",
                    "type": "string",
                    "example": "RUB"
                },
                "percent": {
                    "description": "Значение комиссии (передается, если тип комиссии, \"type\", равен \"percent\" или \"combined\")",
                    "type": "number",
                    "example": 11.99
                },
                "type": {
                    "description": "Тип комиссии:\n* \"percent\" - комиссия в процентах\n* \"fixed\" - фиксированная комиссия\n* \"combined\" - комбинированная комиссия, из процентов и фиксированной суммы\n* \"text\" - текстовая коммиссия, например, \"Взимается провайдером\"",
                    "type": "string",
                    "example": "combined"
                }
            }
        },
        "v1.errorContent": {
            "type": "object",
            "required": [
                "code",
                "description"
            ],
            "properties": {
                "code": {
                    "description": "Код ошибки",
                    "type": "string",
                    "example": "InvalidRequest"
                },
                "description": {
                    "description": "Описание ошибки для разработки",
                    "type": "string",
                    "example": "user_id param is required"
                },
                "message": {
                    "description": "Сообщение об ошибке для клиента",
                    "type": "string",
                    "example": "Internal server error occurred. Please try again later."
                }
            }
        },
        "v1.errorResponse": {
            "type": "object",
            "required": [
                "error",
                "success"
            ],
            "properties": {
                "error": {
                    "description": "Развернутая информация об ошибке",
                    "allOf": [
                        {
                            "$ref": "#/definitions/v1.errorContent"
                        }
                    ]
                },
                "success": {
                    "description": "Результат обработки запроса (всегда false)",
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "v1.favoritesRequest": {
            "type": "object",
            "required": [
                "currency",
                "external_method",
                "external_system",
                "lang_code",
                "session_id",
                "user_id"
            ],
            "properties": {
                "currency": {
                    "description": "Валюта платежного метода платежной системы, которого необходимо занести в \"Избранное\"",
                    "type": "string",
                    "example": "RUB"
                },
                "external_method": {
                    "description": "Внутренний код платежного метода платежной системы, которого необходимо занести в \"Избранное\"",
                    "type": "string",
                    "example": "yookassa_bank_card"
                },
                "external_system": {
                    "description": "Внутренний код платежной системы, платежный метод которой необходимо занести в \"Избранное\"",
                    "type": "string",
                    "example": "yookassa"
                },
                "lang_code": {
                    "description": "Код языка, обозначение по RFC 5646",
                    "type": "string",
                    "example": "en"
                },
                "session_id": {
                    "description": "Идентификатор сессии клиента",
                    "type": "string",
                    "example": "LRXZmXPGusPCfys48LadjFew"
                },
                "user_id": {
                    "description": "Идентификатор клиента",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "v1.favoritesResponse": {
            "type": "object",
            "required": [
                "success"
            ],
            "properties": {
                "success": {
                    "description": "Результат обработки запроса (всегда true)",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "v1.limits": {
            "type": "object",
            "required": [
                "currency",
                "max_amount",
                "min_amount"
            ],
            "properties": {
                "currency": {
                    "description": "Код валюты лимита в соответствии со стандартом ISO 4217",
                    "type": "string",
                    "example": "RUB"
                },
                "max_amount": {
                    "description": "Максимальное значение суммы (в дробных единицах)",
                    "type": "number",
                    "example": 60000
                },
                "min_amount": {
                    "description": "Минимальное значение суммы (в дробных единицах)",
                    "type": "number",
                    "example": 100
                }
            }
        },
        "v1.method": {
            "type": "object",
            "required": [
                "commission",
                "external_method",
                "external_system",
                "favorite",
                "id",
                "limits",
                "name"
            ],
            "properties": {
                "commission": {
                    "description": "Объект, содержащий данные о комиссии",
                    "allOf": [
                        {
                            "$ref": "#/definitions/v1.commission"
                        }
                    ]
                },
                "external_method": {
                    "description": "Внутренний код платежного метода платежной системы, к которой направляется целевой запрос",
                    "type": "string",
                    "example": "yookassa_bank_card"
                },
                "external_system": {
                    "description": "Внутренний код платежной системы, к которой направляется целевой запрос",
                    "type": "string",
                    "example": "yookassa"
                },
                "favorite": {
                    "description": "Флаг о том, что платежная система добавлена в избранное",
                    "type": "boolean",
                    "example": true
                },
                "id": {
                    "description": "Идентификатор ПС из внутреннего справочника",
                    "type": "string",
                    "example": "CARD"
                },
                "limits": {
                    "description": "Массив объектов, содержащих данные о лимитах",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.limits"
                    }
                },
                "name": {
                    "description": "Название платежной системы",
                    "type": "string",
                    "example": "Банковская карта"
                },
                "tools": {
                    "description": "Массив объектов, содержащих данные о сохраненных платежных средствах",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.tool"
                    }
                }
            }
        },
...
        "v1.paymentCreateRequest": {
            "type": "object",
            "required": [
                "amount",
                "currency",
                "external_method",
                "external_system",
                "lang_code",
                "return_urls",
                "session_id",
                "user_id"
            ],
            "properties": {
                "additional_data": {
                    "description": "Дополнительная информация, специфичная для платежной системы, к которой направляется целевой запрос",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "ip": "127.0.0.1",
                        "phone_number": "+71234567890"
                    }
                },
                "amount": {
                    "description": "Сумма платежа в минорных единицах валюты (копейки, центы и т.п.)",
                    "type": "integer",
                    "minimum": 100,
                    "example": 10000
                },
                "currency": {
                    "description": "Валюта платежа в соответствии со стандартом ISO 4217",
                    "type": "string",
                    "example": "RUB"
                },
                "external_method": {
                    "description": "Внутренний код платежного метода платежной системы, к которой направляется целевой запрос",
                    "type": "string",
                    "example": "yookassa_bank_card"
                },
                "external_system": {
                    "description": "Внутренний код платежной системы, к которой направляется целевой запрос",
                    "type": "string",
                    "example": "yookassa"
                },
                "lang_code": {
                    "description": "Код языка, обозначение по RFC 5646",
                    "type": "string",
                    "example": "en"
                },
                "return_urls": {
                    "description": "Объект, содержащий ссылки для возврата клиента для каждого из возможных результатов проведения платежа",
                    "allOf": [
                        {
                            "$ref": "#/definitions/v1.paymentReturnURLs"
                        }
                    ]
                },
                "session_id": {
                    "description": "Идентификатор сессии клиента",
                    "type": "string",
                    "example": "LRXZmXPGusPCfys48LadjFew"
                },
                "tool_id": {
                    "description": "Идентификатор сохраненного платежного средства",
                    "type": "string",
                    "example": "2dc32aa0-000f-5000-8000-16d7bc6cd09f"
                },
                "user_id": {
                    "description": "Идентификатор клиента",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "v1.paymentCreateResponse": {
            "type": "object",
            "required": [
                "operation_id",
                "success",
                "type"
            ],
            "properties": {
                "message": {
                    "description": "Сообщение, которое необходимо показать клиенту",
                    "type": "string",
                    "example": "Баланс пополнен!"
                },
                "operation_id": {
                    "description": "Идентификатор созданного платежа",
                    "type": "integer",
                    "example": 1
                },
                "redirect_url": {
                    "description": "URL платежной страницы, на которую необходимо перенаправить клиента",
                    "type": "string",
                    "example": "https://securepayments.example.com"
                },
                "success": {
                    "description": "Результат обработки запроса (всегда true)",
                    "type": "boolean",
                    "example": true
                },
                "type": {
                    "description": "Тип ответа:\n* Перенаправление клиента на платежную страницу - \"redirect\"\n* Текстовое сообщение - \"message\"",
                    "type": "string",
                    "example": "redirect"
                }
            }
        },
        "v1.paymentMethodsResponse": {
            "type": "object",
            "required": [
                "payment_methods",
                "success"
            ],
            "properties": {
                "payment_methods": {
                    "description": "Массив платежных методов, доступных для пополнения баланса",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.method"
                    }
                },
                "success": {
                    "description": "Результат обработки запроса (всегда true)",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "v1.paymentReturnURLs": {
            "type": "object",
            "required": [
                "common"
            ],
            "properties": {
                "common": {
                    "description": "URL для возврата клиента, используемый когда результат платежа неизвестен или по умолчанию",
                    "type": "string",
                    "example": "https://example.com"
                },
                "fail": {
                    "description": "URL для возврата клиента, используемый при неуспешном осуществлении платежа",
                    "type": "string",
                    "example": "https://example.com/failed"
                },
                "success": {
                    "description": "URL для возврата клиента, используемый при успешном осуществлении платежа",
                    "type": "string",
                    "example": "https://example.com/success"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Секретный ключ",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}